支持 azure_openai
