+++ README.md
@@ -10,7 +10,7 @@
 - 单聊天框的，多会话支持，可切换聊天场景,不丢失对话内容
 - 支持分段极速响应
 - 支持动态修改 prompt ，与预定义 prompt 模板
-- 支持 http/sock5 代理,支持反向域名代理
+- 支持 http/sock5 代理,支持反向域名代理,支持 azure_openai
 - 支持图片消息（答题场景）
 - 支持插件机制（可扩展性👍👍），
   - 可按照规则自行开发接入其他能力（知识库, summary , search...）
