+++ chat/common/openai/openai.go
@@ -2,9 +2,7 @@ package openai
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"io"
 	"net"
 	"net/http"
 	"net/url"
@@ -32,6 +30,8 @@ type ChatModelMessage struct {
 
 type ChatClient struct {
 	APIKey      string  `json:"api_key"`
+	Origin      string  `json:"origin"`
+	Engine      string  `json:"engine"`
 	HttpProxy   string  `json:"http_proxy"`
 	Socks5Proxy string  `json:"socks5_proxy"`
 	Model       string  `json:"model"`
@@ -46,6 +46,18 @@ func NewChatClient(apiKey string) *ChatClient {
 		MaxToken:    MaxToken,
 		Temperature: Temperature,
 	}
+}
+
+// WithOrigin 设置origin
+func (c *ChatClient) WithOrigin(origin string) *ChatClient {
+	c.Origin = origin
+	return c
+}
+
+// WithEngine 设置engine
+func (c *ChatClient) WithEngine(engine string) *ChatClient {
+	c.Engine = engine
+	return c
 }
 
 func (c *ChatClient) WithModel(model string) *ChatClient {
@@ -139,44 +151,11 @@ func (c *ChatClient) Completion(req string) (string, error) {
 	return txt, nil
 }
 
-func (c *ChatClient) Chat(req []ChatModelMessage) (string, error) {
-
-	config := c.buildConfig()
-	cli := copenai.NewClientWithConfig(config)
-
-	// 打印请求信息
-	logx.Info("req: ", req)
-
-	var messages []copenai.ChatCompletionMessage
-
-	for _, message := range req {
-		messages = append(messages, copenai.ChatCompletionMessage{
-			Role:    message.Role,
-			Content: message.Content,
-		})
-	}
-	request := copenai.ChatCompletionRequest{
-		Model:       ChatModel,
-		Messages:    messages,
-		MaxTokens:   c.MaxToken,
-		Temperature: c.Temperature,
-		TopP:        1,
-	}
-	chat, err := cli.CreateChatCompletion(context.Background(), request)
-	if err != nil {
-		fmt.Println("req chat params:", config)
-		return "", err
-	}
-	txt := ""
-	for _, choice := range chat.Choices {
-		txt += choice.Message.Content
-	}
-
-	return txt, nil
-}
-
 func (c *ChatClient) buildConfig() copenai.ClientConfig {
 	config := copenai.DefaultConfig(c.APIKey)
+	if c.Origin == "azure" || c.Origin == "azure_ad" {
+		config = copenai.DefaultAzureConfig(c.APIKey, c.BaseHost, c.Engine)
+	}
 	if c.HttpProxy != "" {
 		proxyUrl, _ := url.Parse(c.HttpProxy)
 		transport := &http.Transport{
@@ -196,95 +175,9 @@ func (c *ChatClient) buildConfig() copenai.ClientConfig {
 		}
 	}
 
-	if c.BaseHost != "" {
+	if c.BaseHost != "" && c.Origin == "open_ai" {
 		// trim last slash
 		config.BaseURL = strings.TrimRight(c.BaseHost, "/") + "/v1"
 	}
 	return config
-}
-
-// ChatStream 数据流式传输
-func (c *ChatClient) ChatStream(req []ChatModelMessage, channel chan string) (string, error) {
-
-	config := c.buildConfig()
-
-	cli := copenai.NewClientWithConfig(config)
-
-	// 打印请求信息
-	logx.Info("req: ", req)
-	first := 0
-	var system ChatModelMessage
-	for i, msg := range req {
-		if msg.Role == "system" {
-			system = msg
-		}
-		if i%2 == 0 {
-			continue
-		}
-		//估算长度
-		if NumTokensFromMessages(req[len(req)-i-1:], ChatModel) < (3900 - c.MaxToken) {
-			first = len(req) - i - 1
-		} else {
-			break
-		}
-	}
-
-	var messages []copenai.ChatCompletionMessage
-
-	if first != 0 {
-		messages = append(messages, copenai.ChatCompletionMessage{
-			Role:    system.Role,
-			Content: system.Content,
-		})
-	}
-
-	for _, message := range req[first:] {
-		messages = append(messages, copenai.ChatCompletionMessage{
-			Role:    message.Role,
-			Content: message.Content,
-		})
-	}
-	request := copenai.ChatCompletionRequest{
-		Model:       ChatModel,
-		Messages:    messages,
-		MaxTokens:   c.MaxToken,
-		Temperature: c.Temperature,
-		TopP:        1,
-	}
-	stream, err := cli.CreateChatCompletionStream(context.Background(), request)
-
-	if err != nil {
-		fmt.Println("req chat stream params:", config)
-		return "", err
-	}
-	defer stream.Close()
-
-	messageText := ""
-	for {
-		response, err := stream.Recv()
-		if errors.Is(err, io.EOF) {
-			logx.Info("Stream finished")
-			return messageText, nil
-		}
-
-		if err != nil {
-			fmt.Printf("Stream error: %v\n", err)
-			close(channel)
-			return messageText, err
-		}
-
-		if len(response.Choices) > 0 {
-			content := response.Choices[0].Delta.Content
-			channel <- content
-			// 如果是对内容的进行补充
-			messageText += content
-			// 结算
-			if response.Choices[0].FinishReason != "" {
-				close(channel)
-				return messageText, nil
-			}
-		}
-
-		logx.Info("Stream response:", response)
-	}
 }
