+++ chat/go.mod
@@ -17,7 +17,7 @@ require (
 	github.com/google/uuid v1.3.0
 	github.com/milvus-io/milvus-sdk-go/v2 v2.2.1
 	github.com/pkg/errors v0.9.1
-	github.com/sashabaranov/go-openai v1.6.1
+	github.com/sashabaranov/go-openai v1.8.0
 	github.com/whyiyhw/go-workwx v0.0.9
 	github.com/zeromicro/go-zero v1.4.4
 	golang.org/x/crypto v0.5.0
